==> R-scripts/08_covariance_messing_around.Rmd <==
---
title: "Covariance messing around"
output:
  html_document: default
  html_notebook: default
  pdf_document: default
---

```{r}
knitr::opts_chunk$set(message = FALSE)
knitr::opts_chunk$set(warning = FALSE)
```

```{r}
library(tidyverse)
library(cowplot)
theme_set(theme_cowplot())
library(roll)
```



#### Invent some TPCs (purple is f(T), orange is g(T))
```{r}
f_curve <- function(x) 0.06*exp(0.09*x)*(1-((x-15)/(34/2))^2)
g_curve <- function(x) 0.01*exp(0.15*x)
p <- ggplot(data = data.frame(x = 0), mapping = aes(x = x)) 

tpcs <- p + 
	stat_function(fun = f_curve, color = "purple", size = 2) +
	stat_function(fun = g_curve, color = "orange", size = 2) +
	xlim(0, 35) + ylim(-0.1, 0.5) + 
	xlab("Temperature (°C)") + ylab("Trait") + geom_hline(yintercept = 0)

tpcs

```





#### make up temperature time series
```{r}
set.seed(101)
time <- seq_along(1:50)
temps <- data.frame(temperature = runif(50, min = 5, max = 33), time = time)

```



#### plot temps over time

```{r}
temp_plot <- temps %>% 
	ggplot(aes(x = time, y = temperature)) + geom_line() +
	ylab("Temperature (°C)") + xlab("Time (days)")
```


#### get trait values

```{r}
f_time_series <- data.frame(time = time, ftrait = sapply(X = temps$temperature, FUN = f_curve), temperature = temps$temperature)
g_time_series <- data.frame(time = time, gtrait = sapply(X = temps$temperature, FUN = g_curve), temperature = temps$temperature)
```

#### plot traits over time
```{r}
both <- left_join(f_time_series, g_time_series) %>% 
	mutate(R_0 = ftrait*gtrait)


trait_plot <- both %>% 
	gather(key = trait_type, value = trait_value, ftrait, gtrait) %>% 
	mutate(trait_type = ifelse(trait_type == "ftrait", "f(T)", "g(T)")) %>% 
	ggplot(aes(x = time, y = trait_value, color = trait_type)) + geom_line(alpha = 0.5) +
	scale_color_manual(values = c("purple", "orange")) +ylab("Trait value") + xlab("Time (days)") +
	theme(legend.position = "top")
```





#### Calculate rolling covariances over the time series with different time slice durations (2, 5, 10, 20 days)


```{r}
traits <- both %>% 
	select(ftrait, gtrait) %>% 
	as.matrix()
results2 <- roll_cov(traits, width = 2, center = FALSE, scale = FALSE)
covariances2 <- as.data.frame(results2) %>% 
	select(contains("f")) %>% 
	rownames_to_column(var = "covariance") %>% 
	filter(covariance == "gtrait") %>% 
	gather(key = time_point, value = covariance) %>% 
	mutate(time_point = str_replace_all(time_point, "[a-z, .]", "")) %>% 
	mutate(time_point = as.numeric(time_point)) %>% 
	rename(covariance2 = covariance)

results5 <- roll_cov(traits, width = 5, center = FALSE, scale = FALSE)
covariances5 <- as.data.frame(results5) %>% 
	select(contains("f")) %>% 
	rownames_to_column(var = "covariance") %>% 
	filter(covariance == "gtrait") %>% 
	gather(key = time_point, value = covariance) %>% 
	mutate(time_point = str_replace_all(time_point, "[a-z, .]", "")) %>% 
	mutate(time_point = as.numeric(time_point)) %>% 
	rename(covariance5 = covariance)

results10 <- roll_cov(traits, width = 10, center = FALSE, scale = FALSE)
covariances10 <- as.data.frame(results10) %>% 
	select(contains("f")) %>% 
	rownames_to_column(var = "covariance") %>% 
	filter(covariance == "gtrait") %>% 
	gather(key = time_point, value = covariance) %>% 
	mutate(time_point = str_replace_all(time_point, "[a-z, .]", "")) %>% 
	mutate(time_point = as.numeric(time_point)) %>% 
	rename(covariance10 = covariance)

results20 <- roll_cov(traits, width = 20, center = FALSE, scale = FALSE)
covariances20 <- as.data.frame(results20) %>% 
	select(contains("f")) %>% 
	rownames_to_column(var = "covariance") %>% 
	filter(covariance == "gtrait") %>% 
	gather(key = time_point, value = covariance) %>% 
	mutate(time_point = str_replace_all(time_point, "[a-z, .]", "")) %>% 
	mutate(time_point = as.numeric(time_point)) %>% 
	rename(covariance20 = covariance)



all_cov <- left_join(covariances2, covariances5) %>% 
	left_join(covariances10) %>% 
	left_join(covariances20) %>% 
	gather(key = slice, value = covariance, 2:5) %>% 
	mutate(slice = str_replace_all(slice, "[a-z, .]", "")) %>% 
	rename(time_slice = slice) %>% 
	mutate(time_slice = as.numeric(time_slice))


cov_plot <- all_cov %>% 
	ggplot(aes(x = time_point, y = covariance, color = factor(time_slice), group = time_slice)) + geom_line(size = 1) +
	 scale_color_viridis_d(name = "Duration of time slice") + geom_hline(yintercept = 0) + ylab("Cov[f(T), g(T)]") +
	xlab("Time (days)") + theme(legend.position = "top")
```

#### Plot all together

```{r}
tpcs
```

```{r, fig.height=12, fig_width = 8 }
all_plots1 <- plot_grid(temp_plot, trait_plot, cov_plot, nrow = 3, ncol = 1)
all_plots1
```

### Now, focusing on the 2 day time slice only, check that `<R0(T)>` =  `<f(T)g(T)>` =  `<f(T)><g(T)>` + `Cov[f(T), g(T)]`
Looks like this isn't working as I expected! i.e. the blue and green lines below should be lining up...Hmmm....

```{r, fig_width = 14}

### get time rolling average for R_0, i.e. <R0(T)> =  <f(T)g(T)>
R_0 = both %>% 
	mutate(R_0 = ftrait*gtrait) %>% 
	select(R_0) %>% 
	as.matrix()
results2_R_0 <- roll_mean(R_0, width = 2)
mean2 <- data.frame(mean_R0 = results2_R_0, complete_obs = TRUE) %>% 
	rownames_to_column(var = "time")


### get time rolling average for f(T), <f(T)>
f_avg <- both %>% 
	select(ftrait) %>% 
	as.matrix()
results2_fmean <- roll_mean(f_avg, width = 2)
meanf2 <- data.frame(mean_f = results2_fmean, complete_obs = TRUE) %>% 
	rownames_to_column(var = "time")

### get time rolling average for g(T), <g(T)>
g_avg <- both %>% 
	select(gtrait) %>% 
	as.matrix()
results2_gmean <- roll_mean(g_avg, width = 2, complete_obs = TRUE)
meang2 <- data.frame(mean_g = results2_gmean) %>% 
	rownames_to_column(var = "time")


### put all the averages together
all_avgs <- left_join(mean2, meanf2) %>% 
	left_join(meang2) %>% 
	mutate(time = as.numeric(time)) %>% 
	left_join(covariances2, by = c("time" = "time_point")) %>% 
	mutate(`<f(T)><g(T)> + Cov[f(T), g(T)]` = (ftrait*gtrait) + covariance2) %>% 
	select(-complete_obs) %>% 
	mutate(`<f(T)><g(T)>` = ftrait*gtrait) %>% 
	gather(key = estimate, value = value, 2:7)

## plot
all_avgs %>% 
	# filter(estimate %in% c("<f(T)><g(T)> + Cov[f(T), g(T)]", "R_0", "<f(T)><g(T)>", "covariance2")) %>% 
	filter(estimate %in% c("<f(T)><g(T)> + Cov[f(T), g(T)]", "R_0")) %>% 
	mutate(estimate = str_replace(estimate, "R_0", "<f(T)g(T)>")) %>% 
	# mutate(estimate = str_replace(estimate, "covariance2", "Cov[f(T), g(T)]")) %>% 
	ggplot(aes(x = value, fill = estimate)) + geom_histogram(alpha = 0.5) 

```

### Now, focusing on the 10 day time slice only, check that `<R0(T)>` =  `<f(T)g(T)>` =  `<f(T)><g(T)>` + `Cov[f(T), g(T)]`
Looks like this isn't working as I expected! i.e. the blue and green lines below should be lining up...Hmmm....

```{r, fig_width = 14}

### get time rolling average for R_0, i.e. <R0(T)> =  <f(T)g(T)>
R_0 = both %>% 
	mutate(R_0 = ftrait*gtrait) %>% 
	select(R_0) %>% 
	as.matrix()
results10_R_0 <- roll_mean(R_0, width = 10)
mean10 <- data.frame(mean_R0 = results10_R_0, complete_obs = TRUE) %>% 
	rownames_to_column(var = "time")


### get time rolling average for f(T), <f(T)>
f_avg <- both %>% 
	select(ftrait) %>% 
	as.matrix()
results10_fmean <- roll_mean(f_avg, width = 10)
meanf10 <- data.frame(mean_f = results10_fmean, complete_obs = TRUE) %>% 
	rownames_to_column(var = "time")

### get time rolling average for g(T), <g(T)>
g_avg <- both %>% 
	select(gtrait) %>% 
	as.matrix()
results10_gmean <- roll_mean(g_avg, width = 10, complete_obs = TRUE)
meang10 <- data.frame(mean_g = results10_gmean) %>% 
	rownames_to_column(var = "time")


### put all the averages together
all_avgs_10 <- left_join(mean10, meanf10) %>% 
	left_join(meang10) %>% 
	mutate(time = as.numeric(time)) %>% 
	left_join(covariances2, by = c("time" = "time_point")) %>% 
	mutate(`<f(T)><g(T)> + Cov[f(T), g(T)]` = (ftrait*gtrait) + covariance2) %>% 
	select(-complete_obs) %>% 
	mutate(`<f(T)><g(T)>` = ftrait*gtrait) %>% 
	gather(key = estimate, value = value, 2:7)

## plot
all_avgs_10 %>% 
	filter(estimate %in% c("<f(T)><g(T)> + Cov[f(T), g(T)]", "R_0")) %>% 
	mutate(estimate = str_replace(estimate, "R_0", "<f(T)g(T)>")) %>% 
	# mutate(estimate = str_replace(estimate, "covariance2", "Cov[f(T), g(T)]")) %>% 
	ggplot(aes(x = value, fill = estimate)) + geom_histogram(alpha = 0.5) 

```

#### Ok let's try something simpler.
Now just split the time series into 20 chunks of 500 time points each. Calculate the covariances and averages in each of those chunks.
```{r, fig.width= 10}


group <-  sort(rep(seq(from = 1, to = 20), times = 500))

all <- both %>% 
	mutate(group = group) %>% 
	mutate(R_0 = ftrait*gtrait) %>% 
	group_by(group) %>% 
	mutate(`<f(T)g(T)>` = mean(R_0)) %>%
	mutate(`<f(T)>`= mean(ftrait)) %>% 
	mutate(`<g(T)>` = mean(gtrait)) %>% 
	mutate(`Cov[f(T), g(T)]` = cov(ftrait, gtrait)) %>% 
	mutate(`<f(T)><g(T)> + Cov[f(T), g(T)]` = (`<f(T)>` * `<g(T)>`) + `Cov[f(T), g(T)]`) %>% 
	gather(key = estimate, value = value, 7:11)

all_wide <- both %>% 
	mutate(group = group) %>% 
	mutate(R_0 = ftrait*gtrait) %>% 
	group_by(group) %>% 
	mutate(`<f(T)g(T)>` = mean(R_0)) %>%
	mutate(`<f(T)>`= mean(ftrait)) %>% 
	mutate(`<g(T)>` = mean(gtrait)) %>% 
	mutate(`Cov[f(T), g(T)]` = cov(ftrait, gtrait)) %>% 
	mutate(`<f(T)><g(T)> + Cov[f(T), g(T)]` = (`<f(T)>` * `<g(T)>`) + `Cov[f(T), g(T)]`)

all_wide %>% 
	rename(`f(T)` = ftrait) %>% 
	rename(`g(T)` = gtrait) %>% 
	ggplot(aes(x = `f(T)`, y = `g(T)`, color = `Cov[f(T), g(T)]`)) + geom_point(alpha = 0.5) +
	scale_color_viridis_c() +
	facet_wrap( ~ group)

all_wide %>% 
	rename(`f(T)` = ftrait) %>% 
	rename(`g(T)` = gtrait) %>% 
	ggplot(aes(x = `f(T)`, y = `g(T)`, color = temperature)) + geom_point(alpha = 0.5) +
	scale_color_viridis_c() +
	facet_wrap( ~ group)
```

```{r}
all %>% 
	ggplot(aes(x = time, y = value, color = estimate)) + geom_line() 

all %>% 
	filter(estimate %in% c("Cov[f(T), g(T)]")) %>% 
	ggplot(aes(x = time, y = value, color = estimate)) + geom_line() 
```


OK this is looking a lot better in terms of matching up
```{r}
all %>% 
	filter(estimate %in% c("<f(T)g(T)>", "<f(T)><g(T)> + Cov[f(T), g(T)]")) %>% 
	ggplot(aes(x = time, y = value, color = estimate)) + geom_line() 

all %>% 
	filter(estimate %in% c("<f(T)g(T)>", "<f(T)><g(T)> + Cov[f(T), g(T)]")) %>% 
	ggplot(aes(x = value, fill = estimate, color = estimate)) + geom_density(alpha = 0.5) 
```


```{r}

 both %>% 
	sample_n(size = 50) %>% 
	mutate(R_0 = ftrait*gtrait) %>% 
	summarise(`<f(T)g(T)>` = mean(R_0),
			  `<f(T)>`= mean(ftrait),
			  `<g(T)>` = mean(gtrait),
			  `Cov[f(T), g(T)]` = cov(ftrait, gtrait)) %>% 
	mutate(`<f(T)><g(T)> + Cov[f(T), g(T)]` = (`<f(T)>` * `<g(T)>`) + `Cov[f(T), g(T)]`) %>%
	mutate(difference = `<f(T)><g(T)> + Cov[f(T), g(T)]` - `<f(T)g(T)>`) 

```
